added apis and changed to _id

+++ user.py
@@ -8,9 +8,11 @@ from flasgger import Swagger
 from auth_token import token_required
 from functools import wraps
 import jwt as jwt_module
-user_routes = Blueprint('users_routes', __name__)
+
 from flask_cors import CORS
 from bson import ObjectId
+
+user_routes = Blueprint('users_routes', __name__)
 
 logger = setup_logger()
 user_collection = UserConfig.get_users_collection()
@@ -26,20 +28,19 @@ def get_endusers_list():
         description: A list of users
       500:
         description: Internal Server Error Occurred
-                   
     """
     try:
         logger.info('Fetching user list')
         users = list(user_collection.find({}))
         for user in users:
-          user['_id'] = str(user['_id'])
-        return jsonify(users),200
+            user['_id'] = str(user['_id'])
+        return jsonify(users), 200
     except Exception as e:
         logger.error('Error fetching user list: %s', str(e), exc_info=True)
         return jsonify({'error': 'Internal server error'}), 500
-    
-@user_routes.route('/users/<string:empid>/access', methods=['GET'])
-def get_access_details(empid):
+
+@user_routes.route('/users/access/<string:empid>', methods=['GET'])
+def get_access_details_by_empid(empid):
     """
     Get enduser access details by empid
     ---
@@ -55,23 +56,27 @@ def get_access_details(empid):
       404:
         description: User not found
       500:
-        description: Internal server error        
+        description: Internal server error
     """
     try:
         logger.info('Fetching user access details for empid: %s', empid)
-        user = user_collection.find_one({'EmpId': empid}, {'_id': 0, 'Access': 1, 'SpaceName': 1})
+        empid=ObjectId(empid)
+        user = user_collection.find_one({'_id': empid}, {'_id': 1,'Access': 1})
+        
         if not user:
             return jsonify({"error": "User not found"}), 404
         
+        user['_id'] = str(user['_id'])
+
         return jsonify(user), 200
     except Exception as e:
         logger.error('Error fetching user access details: %s', str(e), exc_info=True)
         return jsonify({"error": "Internal server error occurred"}), 500
-    
-@user_routes.route('/access_user', methods=['GET'])
-def get_accessed_user():
+
+@user_routes.route('/users_with_access', methods=['GET'])
+def get_users_details_who_has_access():
     """
-    Get details of users with access details
+    Get details of users who has access
     ---
     responses:
       200:
@@ -84,16 +89,19 @@ def get_accessed_user():
     try:
         logger.info('Fetching users with access')
         query = {'Access': {'$ne': []}}
-        users = list(user_collection.find(query, {"_id": 0}))
-        
+        users = list(user_collection.find(query))
+
         if not users:
             return jsonify({"error": "No users found with access"}), 404
         
+        for user in users:
+            user['_id'] = str(user['_id'])
+
         return jsonify(users), 200
     except Exception as e:
         logger.error('Error fetching accessed user list: %s', str(e), exc_info=True)
         return jsonify({'error': 'Internal server error'}), 500
-    
+
 @user_routes.route('/create_user', methods=['POST'])
 @token_required
 def create_users_data(**kwargs):
@@ -101,6 +109,9 @@ def create_users_data(**kwargs):
     Create a user
     ---
     parameters:
+      - name: Authorization
+        in: header
+        required: true
       - name: body
         in: body
         required: true
@@ -127,36 +138,31 @@ def create_users_data(**kwargs):
               type: array
               items:
                 type: string
+            Role:
+              type: string    
     responses:
       201:
         description: User data added successfully
       400:
         description: Error in creating user data
+      403:
+        description: Permission denied  
       409:
         description: Duplicate user ID
+      500:
+        description: Internal server error  
     """
     try:
-        uploader_access = kwargs.get('uploader_access')
-        # cookies = request.cookies
-        # print(f"Received cookies: {request.headers.get('Authorization')}")
-        # print(f"Received cookies: {request.cookies.get('access')}")
-        secret_key = 'St@and@100ardapi@aap100mor#100'
-        token = request.headers.get('Authorization')
-        uploader_data = jwt_module.decode(token, secret_key, algorithms=["HS256"])
-        # print('UPLOADER DATA', uploader_data)
-        if uploader_data['Role'] != 'Admin':
-          return jsonify({'message': 'Permission denied'}), 403
+        uploader_role = kwargs.get('uploader_role')
+        if 'Admin' not in uploader_role:
+            return jsonify({'message': 'Permission denied'}), 403
         logger.info('A new user data added successfully')
 
         json_data = request.get_json()
-        json_data['Role'] = "User"
-        json_data['Access'] = []
-        print("DATA", json_data)
-        user_data = UserModel(**json_data)
-        
+        user_data = UserModel.parse_obj(json_data)
+
         user_data.Id = str(uuid.uuid4())
         result = user_collection.insert_one(user_data.dict())
-        logger.info('User data created successfully')
         return jsonify({'message': 'User data added successfully', 'document_id': str(result.inserted_id)}), 201
     except errors.DuplicateKeyError:
         logger.error('Duplicate user ID detected: %s', user_data.EmpId)
@@ -190,10 +196,18 @@ def update_user_data(empid, **kwargs):
               type: string
             LastName:
               type: string
+            EmpId:
+              type: string
+            Email:
+              type: string    
             EmployeeType:
               type: string
             Contact:
               type: string
+            JobTitle:
+              type: string
+            SpaceName:
+              type: string    
     responses:
       200:
         description: User data updated successfully
@@ -207,31 +221,30 @@ def update_user_data(empid, **kwargs):
         description: Internal server error
     """
     try:
-        # uploader_access = kwargs.get('uploader_access')
-        # if 'Admin' not in uploader_access:
-        #     return jsonify({'message': 'Permission denied'}), 403
+        uploader_role = kwargs.get('uploader_role')
+        if 'Admin' not in uploader_role:
+            return jsonify({'message': 'Permission denied'}), 403
 
-        secret_key = 'St@and@100ardapi@aap100mor#100'
-        token = request.headers.get('Authorization')
-        uploader_data = jwt_module.decode(token, secret_key,algorithms=["HS256"])
-        # print('UPLOADER DATA', uploader_data)
-        if uploader_data['Role'] != 'Admin':
-          return jsonify({'message': 'Permission denied'}), 403
-        logger.info('A new user data added successfully')
         logger.info('Updating user data for EmpId: %s', empid)
         json_data = request.get_json()
+        empid=ObjectId(empid)
 
-        user = user_collection.find_one({'EmpId': empid})
+        user = user_collection.find_one({'_id': empid})
         if not user:
             return jsonify({'error': 'User not found'}), 404
-
+        
         update_data = {
             'FirstName': json_data.get('FirstName'),
             'LastName': json_data.get('LastName'),
+            'EmpId':json_data.get('EmpId'),
+            'Contact': json_data.get('Contact'),
+            'Email':json_data.get('Email'),
+            'JobTitle':json_data.get('JobTitle'),
             'EmployeeType': json_data.get('EmployeeType'),
-            'Contact': json_data.get('Contact')
+            'SpaceName':json_data.get('SpaceName')
+
         }
-        user_collection.update_one({'EmpId': empid}, {'$set': update_data})
+        user_collection.update_one({'_id': empid}, {'$set': update_data})
         return jsonify({'message': 'User data updated successfully'}), 200
     except Exception as e:
         logger.error('Error updating user data: %s', str(e), exc_info=True)
@@ -277,55 +290,31 @@ def update_user_access(empid, **kwargs):
         description: Internal server error
     """
     try:
-        # uploader_access = kwargs.get('uploader_access')
-        # if 'Admin' not in uploader_access:
-        #     return jsonify({'message': 'Permission denied'}), 403
+        uploader_role = kwargs.get('uploader_role')
+        if 'Admin' not in uploader_role:
+            return jsonify({'message': 'Permission denied'}), 403
 
-        secret_key = 'St@and@100ardapi@aap100mor#100'
-        token = request.headers.get('Authorization')
-        uploader_data = jwt_module.decode(token, secret_key,algorithms=["HS256"])
-        # print('UPLOADER DATA', uploader_data)
-        if uploader_data['Role'] != 'Admin':
-          return jsonify({'message': 'Permission denied'}), 403
-        logger.info('A new user data added successfully')
         logger.info('Updating access details for user with EmpId: %s', empid)
         json_data = request.get_json()
+        empid=ObjectId(empid)
 
-        user = user_collection.find_one({'EmpId': empid})
+        user = user_collection.find_one({'_id': empid})
         if not user:
             return jsonify({'error': 'User not found'}), 404
 
         existing_access = user.get('Access', [])
         new_access = json_data.get('Access', [])
-        print("json_data", request.json)
-        print("existing_space_names", existing_access)
-        print("new_space_names", new_access)
+
         updated_access = list(set(existing_access + new_access))
-        print("updated_access", updated_access)
         update_data = {
             'Access': updated_access,
             'SpaceName': json_data.get('SpaceName', [])
         }
-        user_collection.update_one({'EmpId': empid}, {'$set': update_data})
+        user_collection.update_one({'_id': empid}, {'$set': update_data})
         return jsonify({'message': 'User details updated successfully'}), 200
     except Exception as e:
         logger.error('Error updating user details: %s', str(e), exc_info=True)
         return jsonify({'error': 'Internal server error occurred'}), 500
-
-# @user_routes.route('/users/delete/<string:empid>', methods=['DELETE'])
-# def token_required(f):
-#     @wraps(f)
-#     def decorated_function(*args, **kwargs):
-#         # Implement your token validation logic here
-#         token = request.headers.get('Authorization')
-#         if not token:
-#             return jsonify({"message": "Token is missing"}), 401
-#         # Validate the token (this is just a placeholder logic)
-#         # valid_token = validate_token(token)
-#         # if not valid_token:
-#         #     return jsonify({"message": "Invalid token"}), 401
-#         return f(*args, **kwargs)
-#     return decorated_function
 
 @user_routes.route('/users/delete/<string:empid>', methods=['DELETE'])
 @token_required
@@ -358,7 +347,12 @@ def delete_user(empid, **kwargs):
         description: Internal server error
     """
     try:
-        empid = ObjectId(empid)
+        uploader_role = kwargs.get('uploader_role')
+        if 'Admin' not in uploader_role:
+            return jsonify({'Message': 'Permission denied'}), 403
+        logger.info("Deleting EmpId %s", empid)
+        empid=ObjectId(empid)
+
         result = user_collection.delete_one({"_id": empid})
         if result.deleted_count == 1:
             logger.info('User with EmpId %s deleted successfully', empid)
@@ -369,3 +363,59 @@ def delete_user(empid, **kwargs):
     except Exception as e:
         logger.error('An error occurred: %s', str(e), exc_info=True)
         return jsonify({"message": "An error occurred while deleting user data"}), 500
+
+@user_routes.route('/users/delete/access/<string:empid>/<string:access>', methods=['DELETE'])
+@token_required
+def delete_access_from_user(empid, access, **kwargs):
+    """
+    Delete an access from a user's access list
+    ---
+    parameters:
+      - name: Authorization
+        in: header
+        required: true
+      - name: empid
+        in: path
+        description: id of the user whose access needs to be deleted
+        required: true
+        type: string
+      - name: access
+        in: path
+        description: access to be deleted from the user's access list
+        required: true
+        type: string
+    responses:
+      200:
+        description: Access deleted successfully
+      400:
+        description: Bad request - invalid parameters
+      401:
+        description: Unauthorized  
+      404:
+        description: User not found or access not found in the user's access list
+      500:
+        description: Internal server error
+    """
+    try:
+        uploader_role = kwargs.get('uploader_role')
+        if 'Admin' not in uploader_role:
+            return jsonify({'Message': 'Permission denied'}), 403
+        logger.info("Deleting access '%s' from user with EmpId '%s'", access, empid)
+        empid=ObjectId(empid)
+
+        user = user_collection.find_one({'_id': empid})
+        if not user:
+            return jsonify({"error": 'User not found'}), 404
+
+        access_names = user.get('Access', [])
+
+        if access not in access_names:
+            return jsonify({"error": "Access not found in user's access list"}), 404
+
+        access_names.remove(access)
+
+        user_collection.update_one({'_id': empid}, {'$set': {'Access': access_names}})
+        return jsonify({'Message': f"Access '{access}' deleted successfully"}), 200
+    except Exception as e:
+        logger.error('Error deleting access of user: %s', str(e), exc_info=True)
+        return jsonify({'error': 'Internal server error occurred'}), 500
